# backend/docs/openapi.apifox.cn.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Light OSS API",
    "version": "1.0.0",
    "description": "供 APIFOX 导入的 OpenAPI 3.0 JSON。文档中的路径与 HTTP 方法集合与后端实际注册的 Gin 路由保持一致。"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地后端默认地址"
    }
  ],
...
  "paths": {
...
    "/api/v1/buckets": {
      "get": {
        "tags": ["存储桶"],
        "operationId": "listBuckets",
        "summary": "列出存储桶",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BucketSearchQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "存储桶列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出存储桶。错误代码： bucket_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["存储桶"],
        "operationId": "createBucket",
        "summary": "创建存储桶",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BucketCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "存储桶已创建。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求体或存储桶名称无效。错误代码：invalid_request、invalid_bucket_name。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "存储桶已存在。错误代码： bucket_exists。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "创建存储桶失败。错误代码： bucket_create_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "delete": {
        "tags": ["存储桶"],
        "operationId": "deleteBucket",
        "summary": "删除存储桶",
        "description": "删除存储桶及其所有对象、文件夹标记、关联站点和绑定域名。元数据删除成功后执行存储文件清理。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "存储桶已删除。无响应体。"
          },
          "400": {
            "description": "存储桶名称无效。错误代码：invalid_bucket_name。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "存储桶不存在。错误代码：bucket_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "删除存储桶失败。错误代码：bucket_lookup_failed、bucket_delete_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/buckets/{bucket}/objects": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": ["对象"],
        "operationId": "listObjects",
        "summary": "列表对象",
        "description": "对象首先返回最新。next_cursor是相同前缀过滤器的不透明光标。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PrefixQuery"
          },
          {
            "$ref": "#/components/parameters/ObjectListLimitQuery"
          },
          {
            "$ref": "#/components/parameters/CursorQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "对象页面。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称、前缀或光标无效。错误代码： invalid_bucket_name、invalid_prefix、invalid_cursor。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出对象。错误代码： object_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/buckets/{bucket}/objects/{key}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        },
        {
          "$ref": "#/components/parameters/KeyParam"
        }
      ],
      "get": {
        "tags": ["对象"],
        "operationId": "downloadObject",
        "summary": "下载对象",
        "description": "公开对象可匿名下载。私有对象需要 `Authorization: Bearer <token>`，或使用 `expires` 与 `signature` 查询参数的签名 URL。",
        "parameters": [
          {
            "$ref": "#/components/parameters/OptionalAuthorizationHeader"
          },
          {
            "$ref": "#/components/parameters/DownloadExpiresQuery"
          },
          {
            "$ref": "#/components/parameters/DownloadSignatureQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "对象内容。实际媒体类型与对象存储时的 Content-Type 响应头一致。",
            "headers": {
              "Content-Type": {
                "$ref": "#/components/headers/ContentTypeHeader"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLengthHeader"
              },
              "ETag": {
                "$ref": "#/components/headers/ETagHeader"
              },
              "X-Object-Visibility": {
                "$ref": "#/components/headers/ObjectVisibilityHeader"
              },
              "X-Original-Filename": {
                "$ref": "#/components/headers/OriginalFilenameHeader"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称或对象键无效。错误代码：invalid_bucket_name、invalid_object_key。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "私有对象访问未获授权。错误代码：unauthorized、invalid_signature、signature_expired。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到对象。错误代码： object_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法打开对象内容。错误代码： object_lookup_failed, object_open_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": ["对象"],
        "operationId": "headObject",
        "summary": "获取对象标头",
        "description": "仅返回元数据响应头，不返回响应体。公开对象可匿名访问；私有对象需要 `Authorization: Bearer <token>`；HEAD 请求不接受签名 URL。",
        "parameters": [
          {
            "$ref": "#/components/parameters/OptionalAuthorizationHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "仅返回对象响应头，无响应体。",
            "headers": {
              "Content-Type": {
                "$ref": "#/components/headers/ContentTypeHeader"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLengthHeader"
              },
              "ETag": {
                "$ref": "#/components/headers/ETagHeader"
              },
              "X-Object-Visibility": {
                "$ref": "#/components/headers/ObjectVisibilityHeader"
              },
              "X-Original-Filename": {
                "$ref": "#/components/headers/OriginalFilenameHeader"
              }
            }
          },
          "400": {
            "description": "存储桶名称或对象键无效。错误代码：invalid_bucket_name、invalid_object_key。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "访问私有对象需要有效的 Bearer Token。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到对象。错误代码： object_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法打开对象内容。错误代码： object_lookup_failed, object_open_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["对象"],
        "operationId": "uploadObject",
        "summary": "上传对象",
        "description": "请求体会直接以流式方式写入存储。省略时，`X-Object-Visibility` 默认为 `private`，`X-Allow-Overwrite` 默认为 `false`。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UploadVisibilityHeader"
          },
          {
            "$ref": "#/components/parameters/AllowOverwriteHeader"
          },
          {
            "$ref": "#/components/parameters/OriginalFilenameHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "对象已上传。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称、对象键或可见性无效。错误代码： invalid_bucket_name、invalid_object_key、invalid_visibility。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到存储桶。错误代码： bucket_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "对象键已存在且未允许覆盖。错误代码：object_exists。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "请求正文超过配置的上传大小。错误代码： payload_too_large。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "上传失败。错误代码： bucket_lookup_failed、object_lookup_failed、object_store_failed、object_metadata_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["对象"],
        "operationId": "deleteObject",
        "summary": "软删除对象",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "对象已删除。无响应正文。"
          },
          "400": {
            "description": "存储桶名称或对象键无效。错误代码：invalid_bucket_name、invalid_object_key。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到对象。错误代码： object_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法删除对象。错误代码： object_delete_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/objects/visibility/{key}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        },
        {
          "$ref": "#/components/parameters/KeyParam"
        }
      ],
      "patch": {
        "tags": ["对象"],
        "operationId": "updateObjectVisibility",
        "summary": "更新对象可见性",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateObjectVisibilityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "选项卡可见性已更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求体、存储桶名称、对象键或可见性无效。错误代码：invalid_request、invalid_bucket_name、invalid_object_key、invalid_visibility。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到对象。错误代码： object_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "更新对象可见性失败。错误代码： object_update_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sign/download": {
      "post": {
        "tags": ["签名"],
        "operationId": "signDownload",
        "summary": "生成签名下载路径",
        "description": "为 GET 对象下载创建临时签名相对 API 路径。客户端使用当前 API Base URL 补全该路径。当客户端传入 0 或负值时，`expires_in_seconds` 默认为 `300`，且拒绝大于 `86400` 的值。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignDownloadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已签名的相对下载路径。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedDownloadEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求正文、存储桶名称、对象密钥或过期无效。错误代码： invalid_request、invalid_bucket_name、invalid_object_key、invalid_expiry。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "BucketParam": {
        "name": "bucket",
        "in": "path",
        "required": true,
        "description": "存储桶名称。必须匹配 `^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$`。",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
          "minLength": 3,
          "maxLength": 63
        }
      },
      "KeyParam": {
        "name": "key",
        "in": "path",
        "required": true,
        "allowReserved": true,
        "style": "simple",
        "explode": false,
        "description": "对象键。该占位符对应 Gin 的 `/*key` 通配符，可包含 `/`。传参时请使用不带前导斜杠的原始 key 路径。最大长度为 512；如适用，路径遍历和保留文件夹标记名称会被拒绝。",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
...
      "PrefixQuery": {
        "name": "prefix",
        "in": "query",
        "required": false,
        "description": "可选的前缀过滤器。最大长度： 512。文件夹样式前缀通常应以“/”结尾。",
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
...
      "BucketSearchQuery": {
        "name": "search",
        "in": "query",
        "required": false,
        "description": "应用于 bucket 名称的不区分大小写的子字符串筛选器。",
        "schema": {
          "type": "string"
        }
      },
      "ObjectListLimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "页面大小。默认值为20。最大值为100。",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "CursorQuery": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "上一页返回的不透明光标。",
        "schema": {
          "type": "string"
        }
      },
...
      "OptionalAuthorizationHeader": {
        "name": "Authorization",
        "in": "header",
        "required": false,
        "description": "可选的 Bearer Token，格式为 `Bearer <token>`。对于受保护的管理接口是必需的；对于私有对象读取，则按各接口文档说明按需提供。",
        "schema": {
          "type": "string"
        }
      },
      "UploadVisibilityHeader": {
        "name": "X-Object-Visibility",
        "in": "header",
        "required": false,
        "description": "对象可见性。省略时默认为 `private`。",
        "schema": {
          "type": "string",
          "enum": ["public", "private"]
        }
      },
      "AllowOverwriteHeader": {
        "name": "X-Allow-Overwrite",
        "in": "header",
        "required": false,
        "description": "是否允许覆盖同对象键的已存在对象。省略时默认为 `false`。",
        "schema": {
          "type": "string",
          "enum": ["true", "false"]
        }
      },
      "OriginalFilenameHeader": {
        "name": "X-Original-Filename",
        "in": "header",
        "required": false,
        "description": "元数据和下载标头中显示的原始文件名。URL转义的值由后端解码并清理到安全的基名称。",
        "schema": {
          "type": "string",
          "maxLength": 1024
        }
      },
      "DownloadExpiresQuery": {
        "name": "expires",
        "in": "query",
        "required": false,
        "description": "用于私有对象签名下载的 UNIX 时间戳，需与 `signature` 一起使用。",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "DownloadSignatureQuery": {
        "name": "signature",
        "in": "query",
        "required": false,
        "description": "用于私有对象签名下载的签名，需与 `expires` 一起使用。",
        "schema": {
          "type": "string"
        }
      }
    },
    "headers": {
      "ContentTypeHeader": {
        "description": "存储的对象内容类型。",
        "schema": {
          "type": "string"
        }
      },
      "ContentLengthHeader": {
        "description": "存储的对象大小（以字节为单位）。",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 0
        }
      },
      "ETagHeader": {
        "description": "为存储的对象内容生成的实体标记。",
        "schema": {
          "type": "string"
        }
      },
      "ObjectVisibilityHeader": {
        "description": "存储的对象可见性。",
        "schema": {
          "type": "string",
          "enum": ["public", "private"]
        }
      },
      "OriginalFilenameHeader": {
        "description": "已存储原始文件名。",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "SuccessEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": {
            "type": "string",
            "description": "中间件返回的请求关联ID。"
          },
          "data": {
            "type": "object",
            "description": "接口特定的响应数据。"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "BucketCreateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
            "minLength": 3,
            "maxLength": 63,
            "description": "存储桶名称。仅允许小写字母、数字、点和连字符。"
          }
        }
      },
...
      "UpdateObjectVisibilityRequest": {
        "type": "object",
        "required": ["visibility"],
        "properties": {
          "visibility": {
            "type": "string",
            "enum": ["public", "private"]
          }
        }
      },
...
      "BucketEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/Bucket"
              }
            }
          }
        ]
      },
      "BucketListEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/BucketList"
              }
            }
          }
        ]
      },
...
      "ObjectEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        ]
      },
      "ObjectListEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/ObjectList"
              }
            }
          }
        ]
      },
...
    }
  }
}